+++ src/main/kotlin/blackjack/model/Player.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.model
 
 class Player(name: String) : Gamer(name)
