+++ src/main/kotlin/blackjack/model/Shape.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 enum class Shape(val symbol: String) {
     SPADE("스페이드"),
