+++ src/main/kotlin/blackjack/view/InputView.kt
@@ -1,4 +1,6 @@
-package blackjack
+package blackjack.view
+
+import blackjack.model.Gamer
 
 object InputView {
