+++ src/test/kotlin/blackjack/model/CardTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -7,7 +7,8 @@ class CardTest {
 
     @Test
     fun `카드 만들기`() {
-        val card = Card(Denomination.ACE to Shape.CLUB)
+        val card =
+            Card(Denomination.ACE to Shape.CLUB)
 
         assertThat(card.toString()).isEqualTo("A클로버")
     }
