+++ src/test/kotlin/blackjack/model/PointTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -7,7 +7,9 @@ class PointTest {
 
     @Test
     fun `첫 장이 ACE 일 때 ACE 추가 점수 획득`() {
-        val pointIfAceFirst = Point.calculateIfAceFirst(Point(Denomination.ACE.point))
+        val pointIfAceFirst = Point.calculateIfAceFirst(
+            Point(Denomination.ACE.point)
+        )
 
         assertThat(pointIfAceFirst).isEqualTo(Point(11))
     }
