+++ src/test/kotlin/blackjack/model/RevenueTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
